Refactor connection flow to use dedicated Connect page
Move connection logic and form from Index to new Connect page. Update all redirects for disconnected users to point to /Connect. Remove connection properties and logic from IndexModel. Add Connect page for handling connection input, validation, and redirection. Improves separation of concerns and user experience.
